Moved some more of the logic that was in the `TestRunner` control out into smaller ones
