+++ Jantimizer/ExperimentSuite/UserControls/HistogramGeneratorControl.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="ExperimentSuite.UserControls.ReportMaker"
+﻿<UserControl x:Class="ExperimentSuite.UserControls.HistogramGeneratorControl"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
@@ -6,7 +6,8 @@
              xmlns:local="clr-namespace:ExperimentSuite.UserControls"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800">
-    <DataGrid x:Name="MainGrid">
-
-    </DataGrid>
+    <StackPanel>
+        <Label x:Name="HistogramProgressLabel" Content="Histogram Progress (0/0)"/>
+        <ProgressBar x:Name="HistogramProgressBar" Height="20"/>
+    </StackPanel>
 </UserControl>
