+++ Jantimizer/ExperimentSuite/UserControls/SQLFileProgressControl.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="ExperimentSuite.UserControls.ReportMaker"
+﻿<UserControl x:Class="ExperimentSuite.UserControls.SQLFileProgressControl"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
@@ -6,7 +6,9 @@
              xmlns:local="clr-namespace:ExperimentSuite.UserControls"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800">
-    <DataGrid x:Name="MainGrid">
-
-    </DataGrid>
+    <StackPanel>
+        <Label x:Name="SQLProgressLabel" Content="SQL Progress (0/0)"/>
+        <Label x:Name="CurrentSqlFileLabels" Content="File: " FontSize="10"/>
+        <ProgressBar x:Name="SQLProgressBar" Height="20" ValueChanged="SQLProgressBar_ValueChanged"/>
+    </StackPanel>
 </UserControl>
